Squeeze first dimension of feature_data if len=1
I propose to squeeze the first dimension of feature_data when it is of length 1, as it needs to match the dimensions of the mask layers.

For certain eopatches which have a different number of dates for the IS_VALID mask than for the data FEATURES, this is necessary in order for the processing not to encounter errors. That said, I would be interested to know what the origin of the problem is (i.e. a number of IS_VALID mask inferior to the number of FEATURES or BANDS in the eopatch).